+++ savant_core/savant_core_py/src/nvinfer/batching_operator.rs
@@ -278,7 +278,7 @@ pub struct PyOperatorElementOutput {
     slot_idx: usize,
     element_idx: usize,
     roi_id: Option<i64>,
-    slot_number: u32,
+    slot_number: i64,
     num_tensors: usize,
 }
 
@@ -304,7 +304,7 @@ impl PyOperatorElementOutput {
 
     /// DeepStream surface slot index (``NvDsFrameMeta.batch_id``).
     #[getter]
-    fn slot_number(&self) -> u32 {
+    fn slot_number(&self) -> i64 {
         self.slot_number
     }
